+++ src/main/java/elocindev/animation_overhaul/mixin/PlayerEntityRendererMixin.java
@@ -41,6 +41,7 @@ public abstract class PlayerEntityRendererMixin
         MultiBufferSource vertexConsumerProvider, int i, CallbackInfo ci) {
 
         Minecraft mc = Minecraft.getInstance();
+        if (mc.gameRenderer == null) return;
 
         if (!mc.gameRenderer.getMainCamera().isDetached()
                 && player == mc.player) {