+++ changelog.md
@@ -1,3 +1,9 @@
+## 1.2.0
+- Config is now moved to the animation_overhaul folder, and the animations config is now animations.json5
+- Outdated configs will now be updated
+- Added drinking animation
+- Added slow falling animation
+
 ## 1.1.0
 - 1.19.2 support for both fabric and forge
 - Fixed crashes on forge 1.20.1